Engine: With VPtrList crashes are gone, but still a task-list to do - Some of NPC settings are seems a bit damaged and doesn't fits into working result. - Possible danger places are still be, still need to test them - Some places are still use Qt types which are must be replaced - Replace UrlOpener with own class - Lots of signed/unsigned comparisons must be fixed - Some cases may be unbuildable on other platforms - etc.

+++ Engine/scenes/level/lvl_player_switch.cpp
@@ -77,7 +77,7 @@ void CharacterSwitcherEngine::refreshState()
 {
     if(m_scene)
     {
-        for(int i = 0; i < m_switches.size(); i++)
+        for(size_t i = 0; i < m_switches.size(); i++)
         {
             CharacterSwitchState &sw = m_switches[i];
             unsigned long needed = static_cast<unsigned long>(sw.CharacterID);
@@ -104,7 +104,7 @@ void CharacterSwitcherEngine::refreshState()
             }
         }
 
-        for(int i = 0; i < m_blocks.size(); i++)
+        for(size_t i = 0; i < m_blocks.size(); i++)
         {
             CharacterBlockState &blk = m_blocks[i];
             unsigned long needed = static_cast<unsigned long>(blk.CharacterID);
